Support for parameters without table added

+++ powerbi-desktop-parameter-edit.py
@@ -5,7 +5,8 @@ import re
 
 parameters = {
     "DateParameterDemo": "#datetime(2025, 5, 17, 0, 0, 0)",
-    "TextParameterDemo": '"B"'
+    "TextParameterDemo": '"B"',
+    "IsolatedParameterDemo": '"demo"'
 }
 
 project = "c:\\temp\\sample\\parameterized report.pbip"
@@ -29,11 +30,20 @@ if os.path.exists(abf_file):
 
 for parameter, value in parameters.items():
     parameter_file = os.path.join(destination_folder, f"definition\\tables\\{parameter}.tmdl")
-    with open(parameter_file, 'r', encoding='utf-8') as file:
-        content = file.read()
-    content = re.sub(r'source\s*=\s*.+meta', f'source={value} meta', content)
-    with open(parameter_file, 'w', encoding='utf-8') as file:
-        file.write(content)
+    if os.path.exists(parameter_file):
+        with open(parameter_file, 'r', encoding='utf-8') as file:
+            content = file.read()
+        content = re.sub(r'source\s*=\s*.+meta', f'source={value} meta', content)
+        with open(parameter_file, 'w', encoding='utf-8') as file:
+            file.write(content)
+    else:
+        expressions_file = os.path.join(destination_folder, f"definition\\expressions.tmdl")
+        with open(expressions_file, 'r', encoding='utf-8') as file:
+            content = file.read()
+        content = re.sub(rf'expression\s{re.escape(parameter)}\s*=\s*.+meta', f'expression {parameter}={value} meta', content)
+        with open(expressions_file, 'w', encoding='utf-8') as file:
+            file.write(content)
+
 
 pbip_file = os.path.join(stage, os.path.basename(project))
 subprocess.run(['start', '', pbip_file], shell=True, check=True)
